+++ src/exceptions.rs
@@ -9,4 +9,5 @@ pub enum Exception {
     InvalidDataType,
     NativeFault,
     IncorrectRegType,
+    HeapSegmFault
 }